requirements.txt
flask==2.0.1
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
python-dotenv==0.19.0 
wheel